backend/requirements.txt
Django==5.2.8
djangorestframework
psycopg2-binary
dj-database-url
python-dotenv
twilio
requests
gunicorn
whitenoise>=6.6.0
numpy
azure-storage-blob
google-generativeai>=0.8.0
django-cors-headers